@@ -1,19 +1,17 @@
 name: notus
 description: Platform-agnostic rich text document model based on Delta format and used in Zefyr editor.
-version: 0.1.5
+version: 1.0.0-dev.8.0
 author: Anatoly Pulyaevskiy <anatoly.pulyaevskiy@gmail.com>
 homepage: https://github.com/memspace/zefyr
 
 environment:
-  sdk: '>=2.2.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  collection: ^1.14.6
-  meta: ^1.1.0
-  quill_delta: ^1.0.0
-  quiver_hashcode: ^2.0.0
+  collection: ^1.15.0
+  coverage: ^1.0.3
+  quill_delta: ^3.0.0-nullsafety.1
+  quiver: ^3.0.1
 
 dev_dependencies:
-  pedantic: ^1.0.0
-  #test: ^1.10.0
-  #test_coverage: ^0.4.0
+  test: ^1.10.0
