@@ -1,27 +1,22 @@
 name: zefyr
 description: Clean, minimalistic and collaboration-ready rich text editor for Flutter.
-version: 0.12.0
+version: 1.0.0-dev.2.0
 author: Anatoly Pulyaevskiy <anatoly.pulyaevskiy@gmail.com>
 homepage: https://github.com/memspace/zefyr
 
 environment:
-  sdk: '>=2.2.2 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   collection: ^1.14.6
-  url_launcher: ^6.0.12
-  quill_delta: ^1.0.0
-  notus:
-    git:
-      url: git://github.com/churchnotesapp/zefyr-master.git
-      path: packages/notus/
-      ref: bdcf3b824d7205aa3f6dce5c6d00a55a8495201c
-  meta: ^1.1.0
-  quiver_hashcode: ^2.0.0
+  url_launcher: ^6.0.9
+  quill_delta: ^3.0.0-nullsafety.1
+  notus: ^1.0.0-dev.8.0
+  characters: ^1.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.0.0
+  flutter_lints: ^1.0.4
