@@ -1,10 +1,10 @@
 // Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:quill_delta/quill_delta.dart';
-import 'package:zefyr/src/widgets/selection.dart';
 import 'package:zefyr/zefyr.dart';
 
 var delta = Delta()..insert('This House Is A Circus\n');
@@ -17,12 +17,11 @@ class EditorSandBox {
   final Widget widget;
 
   factory EditorSandBox({
-    @required WidgetTester tester,
-    FocusNode focusNode,
-    NotusDocument document,
-    ZefyrThemeData theme,
+    required WidgetTester tester,
+    FocusNode? focusNode,
+    NotusDocument? document,
+    ZefyrThemeData? theme,
     bool autofocus = false,
-    ZefyrImageDelegate imageDelegate,
   }) {
     focusNode ??= FocusNode();
     document ??= NotusDocument.fromDelta(delta);
@@ -32,14 +31,13 @@ class EditorSandBox {
       controller: controller,
       focusNode: focusNode,
       autofocus: autofocus,
-      imageDelegate: imageDelegate,
     );
 
     if (theme != null) {
       widget = ZefyrTheme(data: theme, child: widget);
     }
     widget = MaterialApp(
-      home: ZefyrScaffold(child: widget),
+      home: widget,
     );
 
     return EditorSandBox._(tester, focusNode, document, controller, widget);
@@ -55,7 +53,7 @@ class EditorSandBox {
     return tester.pumpAndSettle();
   }
 
-  Future<void> updateSelection({int base, int extent}) {
+  Future<void> updateSelection({required int base, required int extent}) {
     controller.updateSelection(
       TextSelection(baseOffset: base, extentOffset: extent),
     );
@@ -71,10 +69,11 @@ class EditorSandBox {
 
   Future<void> pump() async {
     await tester.pumpWidget(widget);
+    await tester.pumpAndSettle();
   }
 
   Future<void> tap() async {
-    await tester.tap(find.byType(ZefyrParagraph).first);
+    await tester.tap(find.byType(RawEditor).first);
     await tester.pumpAndSettle();
     expect(focusNode.hasFocus, isTrue);
   }
@@ -89,46 +88,38 @@ class EditorSandBox {
   }
 
   Future<void> tapButtonWithIcon(IconData icon) async {
-    await tester.tap(find.widgetWithIcon(ZefyrButton, icon));
+    await tester.tap(find.widgetWithIcon(RawMaterialButton, icon));
     await tester.pumpAndSettle();
   }
 
   Future<void> tapButtonWithText(String text) async {
-    await tester.tap(find.widgetWithText(ZefyrButton, text));
+    await tester.tap(find.widgetWithText(RawMaterialButton, text));
     await tester.pumpAndSettle();
   }
 
-  RawZefyrButton findButtonWithIcon(IconData icon) {
-    final button = tester.widget(find.widgetWithIcon(RawZefyrButton, icon))
-        as RawZefyrButton;
+  RawMaterialButton findButtonWithIcon(IconData icon) {
+    final button = tester.widget(find.widgetWithIcon(RawMaterialButton, icon))
+        as RawMaterialButton;
     return button;
   }
 
-  RawZefyrButton findButtonWithText(String text) {
-    final button = tester.widget(find.widgetWithText(RawZefyrButton, text))
-        as RawZefyrButton;
+  RawMaterialButton findButtonWithText(String text) {
+    final button = tester.widget(find.widgetWithText(RawMaterialButton, text))
+        as RawMaterialButton;
     return button;
-  }
-
-  Finder findSelectionHandle() {
-    return find.descendant(
-        of: find.byType(SelectionHandleDriver),
-        matching: find.byType(GestureDetector));
   }
 }
 
 class _ZefyrSandbox extends StatefulWidget {
   const _ZefyrSandbox({
-    Key key,
-    @required this.controller,
-    @required this.focusNode,
+    Key? key,
+    required this.controller,
+    required this.focusNode,
     this.autofocus = false,
-    this.imageDelegate,
   }) : super(key: key);
   final ZefyrController controller;
   final FocusNode focusNode;
   final bool autofocus;
-  final ZefyrImageDelegate imageDelegate;
 
   @override
   _ZefyrSandboxState createState() => _ZefyrSandboxState();
@@ -139,12 +130,11 @@ class _ZefyrSandboxState extends State<_ZefyrSandbox> {
 
   @override
   Widget build(BuildContext context) {
-    return ZefyrEditor(
+    return ZefyrField(
       controller: widget.controller,
       focusNode: widget.focusNode,
-      mode: _enabled ? ZefyrMode.edit : ZefyrMode.view,
+      readOnly: !_enabled,
       autofocus: widget.autofocus,
-      imageDelegate: widget.imageDelegate,
     );
   }
 
@@ -152,96 +142,5 @@ class _ZefyrSandboxState extends State<_ZefyrSandbox> {
     setState(() {
       _enabled = false;
     });
-  }
-}
-
-class MultiEditorSandbox {
-  final WidgetTester tester;
-  final Key firstEditorKey;
-  final Key secondEditorKey;
-  final FocusNode firstFocusNode;
-  final FocusNode secondFocusNode;
-  final Widget widget;
-
-  factory MultiEditorSandbox({@required WidgetTester tester}) {
-    final firstEditorKey = UniqueKey();
-    final secondEditorKey = UniqueKey();
-    final firstFocusNode = FocusNode();
-    final secondFocusNode = FocusNode();
-    Widget first = _ZefyrSandbox(
-      key: firstEditorKey,
-      controller: ZefyrController(NotusDocument.fromDelta(delta)),
-      focusNode: firstFocusNode,
-    );
-    Widget second = _ZefyrSandbox(
-      key: secondEditorKey,
-      controller: ZefyrController(NotusDocument.fromDelta(delta)),
-      focusNode: secondFocusNode,
-    );
-
-    Widget widget = MaterialApp(
-      home: Scaffold(
-        body: ZefyrScaffold(
-          child: Column(
-            children: <Widget>[
-              SizedBox(height: 100, child: first),
-              SizedBox(height: 10),
-              SizedBox(height: 100, child: second),
-            ],
-          ),
-        ),
-      ),
-    );
-
-    return MultiEditorSandbox._(
-      tester: tester,
-      widget: widget,
-      firstEditorKey: firstEditorKey,
-      secondEditorKey: secondEditorKey,
-      firstFocusNode: firstFocusNode,
-      secondFocusNode: secondFocusNode,
-    );
-  }
-
-  MultiEditorSandbox._({
-    @required this.tester,
-    @required this.widget,
-    @required this.firstEditorKey,
-    @required this.secondEditorKey,
-    @required this.firstFocusNode,
-    @required this.secondFocusNode,
-  });
-
-  Future<void> pump() async {
-    await tester.pumpWidget(widget);
-  }
-
-  Future<void> tapFirstEditor() async {
-    await tester.tap(find.byKey(firstEditorKey).first);
-    await tester.pumpAndSettle();
-  }
-
-  Future<void> tapSecondEditor() async {
-    await tester.tap(find.byKey(secondEditorKey).first);
-    await tester.pumpAndSettle();
-  }
-
-  ZefyrEditor findFirstEditor() {
-    return tester.widget(find.descendant(
-      of: find.byKey(firstEditorKey),
-      matching: find.byType(ZefyrEditor),
-    ));
-  }
-
-  ZefyrEditor findSecondEditor() {
-    return tester.widget(find.descendant(
-      of: find.byKey(secondEditorKey),
-      matching: find.byType(ZefyrEditor),
-    ));
-  }
-
-  Future<void> tapButtonWithIcon(IconData icon) async {
-    await tester.tap(find.widgetWithIcon(ZefyrButton, icon));
-    await tester.pumpAndSettle();
   }
 }
