@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 import 'package:collection/collection.dart';
-import 'package:quiver_hashcode/hashcode.dart';
+import 'package:quiver/core.dart';
 
 /// Scope of a style attribute, defines context in which an attribute can be
 /// applied.
@@ -42,7 +42,7 @@ abstract class NotusAttributeBuilder<T> implements NotusAttributeKey<T> {
   final String key;
   final NotusAttributeScope scope;
   NotusAttribute<T> get unset => NotusAttribute<T>._(key, scope, null);
-  NotusAttribute<T> withValue(T value) =>
+  NotusAttribute<T> withValue(T? value) =>
       NotusAttribute<T>._(key, scope, value);
 }
 
@@ -67,6 +67,8 @@ abstract class NotusAttributeBuilder<T> implements NotusAttributeKey<T> {
 ///
 ///   * [NotusAttribute.bold]
 ///   * [NotusAttribute.italic]
+///   * [NotusAttribute.underline]
+///   * [NotusAttribute.strikethrough]
 ///   * [NotusAttribute.link]
 ///   * [NotusAttribute.heading]
 ///   * [NotusAttribute.block]
@@ -75,10 +77,10 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
     NotusAttribute.bold.key: NotusAttribute.bold,
     NotusAttribute.italic.key: NotusAttribute.italic,
     NotusAttribute.underline.key: NotusAttribute.underline,
+    NotusAttribute.strikethrough.key: NotusAttribute.strikethrough,
     NotusAttribute.link.key: NotusAttribute.link,
     NotusAttribute.heading.key: NotusAttribute.heading,
     NotusAttribute.block.key: NotusAttribute.block,
-    NotusAttribute.embed.key: NotusAttribute.embed,
   };
 
   // Inline attributes
@@ -91,6 +93,9 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
 
   /// Underline style attribute.
   static const underline = _UnderlineAttribute();
+
+  /// Strikethrough style attribute.
+  static const strikethrough = _StrikethroughAttribute();
 
   /// Link style attribute.
   // ignore: const_eval_throws_exception
@@ -127,16 +132,12 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   /// Alias for [NotusAttribute.block.code].
   static NotusAttribute<String> get code => block.code;
 
-  /// Embed style attribute.
-  // ignore: const_eval_throws_exception
-  static const embed = EmbedAttributeBuilder._();
-
   static NotusAttribute _fromKeyValue(String key, dynamic value) {
     if (!_registry.containsKey(key)) {
       throw ArgumentError.value(
           key, 'No attribute with key "$key" registered.');
     }
-    final builder = _registry[key];
+    final builder = _registry[key]!;
     return builder.withValue(value);
   }
 
@@ -158,7 +159,7 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   ///
   /// See also [unset], [NotusStyle.merge] and [NotusStyle.put]
   /// for details.
-  final T value;
+  final T? value;
 
   /// Returns special "unset" version of this attribute.
   ///
@@ -176,17 +177,14 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   bool get isInline => scope == NotusAttributeScope.inline;
 
   @override
-  NotusAttribute<T> withValue(T value) =>
+  NotusAttribute<T> withValue(T? value) =>
       NotusAttribute<T>._(key, scope, value);
 
   @override
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other is! NotusAttribute<T>) return false;
-    NotusAttribute<T> typedOther = other;
-    return key == typedOther.key &&
-        scope == typedOther.scope &&
-        value == typedOther.value;
+    return key == other.key && scope == other.scope && value == other.value;
   }
 
   @override
@@ -204,7 +202,7 @@ class NotusStyle {
 
   final Map<String, NotusAttribute> _data;
 
-  static NotusStyle fromJson(Map<String, dynamic> data) {
+  static NotusStyle fromJson(Map<String, dynamic>? data) {
     if (data == null) return NotusStyle();
 
     final result = data.map((String key, dynamic value) {
@@ -241,16 +239,15 @@ class NotusStyle {
   /// Returns `true` if this set contains attribute with the same value as
   /// [attribute].
   bool containsSame(NotusAttribute attribute) {
-    assert(attribute != null);
     return get<dynamic>(attribute) == attribute;
   }
 
   /// Returns value of specified attribute [key] in this set.
-  T value<T>(NotusAttributeKey<T> key) => get(key).value;
+  T? value<T>(NotusAttributeKey<T> key) => get(key)?.value;
 
   /// Returns [NotusAttribute] from this set by specified [key].
-  NotusAttribute<T> get<T>(NotusAttributeKey<T> key) =>
-      _data[key.key] as NotusAttribute<T>;
+  NotusAttribute<T>? get<T>(NotusAttributeKey<T> key) =>
+      _data[key.key] as NotusAttribute<T>?;
 
   /// Returns collection of all attribute keys in this set.
   Iterable<String> get keys => _data.keys;
@@ -302,7 +299,7 @@ class NotusStyle {
   }
 
   /// Returns JSON-serializable representation of this style.
-  Map<String, dynamic> toJson() => _data.isEmpty
+  Map<String, dynamic>? toJson() => _data.isEmpty
       ? null
       : _data.map<String, dynamic>((String _, NotusAttribute value) =>
           MapEntry<String, dynamic>(value.key, value.value));
@@ -311,9 +308,8 @@ class NotusStyle {
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other is! NotusStyle) return false;
-    NotusStyle typedOther = other;
     final eq = const MapEquality<String, NotusAttribute>();
-    return eq.equals(_data, typedOther._data);
+    return eq.equals(_data, other._data);
   }
 
   @override
@@ -339,6 +335,12 @@ class _ItalicAttribute extends NotusAttribute<bool> {
 /// Applies underline style to a text segment.
 class _UnderlineAttribute extends NotusAttribute<bool> {
   const _UnderlineAttribute() : super._('u', NotusAttributeScope.inline, true);
+}
+
+/// Applies strikethrough style to a text segment.
+class _StrikethroughAttribute extends NotusAttribute<bool> {
+  const _StrikethroughAttribute()
+      : super._('s', NotusAttributeScope.inline, true);
 }
 
 /// Builder for link attribute values.
@@ -396,76 +398,4 @@ class BlockAttributeBuilder extends NotusAttributeBuilder<String> {
   /// Formats a block of lines as a quote.
   NotusAttribute<String> get quote =>
       NotusAttribute<String>._(key, scope, 'quote');
-}
-
-class EmbedAttributeBuilder
-    extends NotusAttributeBuilder<Map<String, dynamic>> {
-  const EmbedAttributeBuilder._()
-      : super._(EmbedAttribute._kEmbed, NotusAttributeScope.inline);
-
-  NotusAttribute<Map<String, dynamic>> get horizontalRule =>
-      EmbedAttribute.horizontalRule();
-
-  NotusAttribute<Map<String, dynamic>> image(String source) =>
-      EmbedAttribute.image(source);
-
-  @override
-  NotusAttribute<Map<String, dynamic>> get unset => EmbedAttribute._(null);
-
-  @override
-  NotusAttribute<Map<String, dynamic>> withValue(Map<String, dynamic> value) =>
-      EmbedAttribute._(value);
-}
-
-/// Type of embedded content.
-enum EmbedType { horizontalRule, image }
-
-class EmbedAttribute extends NotusAttribute<Map<String, dynamic>> {
-  static const _kValueEquality = MapEquality<String, dynamic>();
-  static const _kEmbed = 'embed';
-  static const _kHorizontalRuleEmbed = 'hr';
-  static const _kImageEmbed = 'image';
-
-  EmbedAttribute._(Map<String, dynamic> value)
-      : super._(_kEmbed, NotusAttributeScope.inline, value);
-
-  EmbedAttribute.horizontalRule()
-      : this._(<String, dynamic>{'type': _kHorizontalRuleEmbed});
-
-  EmbedAttribute.image(String source)
-      : this._(<String, dynamic>{'type': _kImageEmbed, 'source': source});
-
-  /// Type of this embed.
-  EmbedType get type {
-    if (value['type'] == _kHorizontalRuleEmbed) return EmbedType.horizontalRule;
-    if (value['type'] == _kImageEmbed) return EmbedType.image;
-    assert(false, 'Unknown embed attribute value $value.');
-    return null;
-  }
-
-  @override
-  NotusAttribute<Map<String, dynamic>> get unset => EmbedAttribute._(null);
-
-  @override
-  bool operator ==(other) {
-    if (identical(this, other)) return true;
-    if (other is! EmbedAttribute) return false;
-    EmbedAttribute typedOther = other;
-    return key == typedOther.key &&
-        scope == typedOther.scope &&
-        _kValueEquality.equals(value, typedOther.value);
-  }
-
-  @override
-  int get hashCode {
-    final objects = [key, scope];
-    if (value != null) {
-      final valueHashes =
-          value.entries.map((entry) => hash2(entry.key, entry.value));
-      objects.addAll(valueHashes);
-    } else {
-      objects.add(value);
-    }
-    return hashObjects(objects);
-  }
 }
