@@ -6,6 +6,7 @@ import 'dart:math' as math;
 import 'package:quill_delta/quill_delta.dart';
 
 import 'attributes.dart';
+import 'embeds.dart';
 import 'line.dart';
 import 'node.dart';
 
@@ -13,29 +14,25 @@ import 'node.dart';
 abstract class LeafNode extends Node
     with StyledNodeMixin
     implements StyledNode {
-  /// Creates a new [LeafNode] with specified [value].
-  LeafNode._([String value = ''])
-      : assert(value != null && !value.contains('\n')),
-        _value = value;
+  /// Creates a new [LeafNode] with specified [data].
+  LeafNode._(Object data) : _value = data;
 
-  factory LeafNode([String value = '']) {
+  factory LeafNode(Object data) {
     LeafNode node;
-    if (value == kZeroWidthSpace) {
-      // Zero-width space is reserved for embed nodes.
-      node = EmbedNode();
+    if (data is EmbeddableObject) {
+      node = EmbedNode(data);
     } else {
-      assert(
-          !value.contains(kZeroWidthSpace),
-          'Zero-width space is reserved for embed leaf nodes and cannot be used '
-          'inside regular text nodes.');
-      node = TextNode(value);
+      final text = data as String;
+      assert(text.isNotEmpty);
+      node = TextNode(text);
     }
     return node;
   }
 
-  /// Plain-text value of this node.
-  String get value => _value;
-  String _value;
+  /// Contents of this node, either a String if this is a [TextNode] or an
+  /// [EmbeddableObject] if this is an [EmbedNode].
+  Object get value => _value;
+  Object _value;
 
   /// Splits this leaf node at [index] and returns new node.
   ///
@@ -48,18 +45,30 @@ abstract class LeafNode extends Node
   ///
   /// In case a new node is actually split from this one, it inherits this
   /// node's style.
-  LeafNode splitAt(int index) {
+  LeafNode? splitAt(int index) {
     assert(index >= 0 && index <= length);
     if (index == 0) return this;
     if (index == length && isLast) return null;
     if (index == length && !isLast) return next as LeafNode;
 
-    final text = _value;
-    _value = text.substring(0, index);
-    final split = LeafNode(text.substring(index));
-    split.applyStyle(style);
-    insertAfter(split);
-    return split;
+    if (this is TextNode) {
+      final text = _value as String;
+      _value = text.substring(0, index);
+      final split = LeafNode(text.substring(index));
+      split.applyStyle(style);
+      insertAfter(split);
+      return split;
+    } else {
+      // This is an EmbedNode which cannot be split and has length of 1.
+      // Technically this branch is unreachable because we've already checked
+      // edge scenarios when index is either at 0 (start) or 1 (end) of this
+      // node above, e.g. for embed nodes:
+      //   * calling splitAt(0) returns the embed node itself;
+      //   * calling splitAt(1) depends on whether this node is the last one:
+      //     - if last then `null` is returned as there is nothing to split
+      //     - otherwise next node is returned
+      throw StateError('Unreachable.');
+    }
   }
 
   /// Cuts a leaf node from [index] to the end of this node and returns new node
@@ -67,7 +76,7 @@ abstract class LeafNode extends Node
   ///
   /// Splitting logic is identical to one described in [splitAt], meaning this
   /// method may return `null`.
-  LeafNode cutAt(int index) {
+  LeafNode? cutAt(int index) {
     assert(index >= 0 && index <= length);
     final cut = splitAt(index);
     cut?.unlink();
@@ -79,7 +88,7 @@ abstract class LeafNode extends Node
   /// Splitting logic is identical to one described in [splitAt], with one
   /// exception that it is required for [index] to always be less than this
   /// node's length. As a result this method always returns a [LeafNode]
-  /// instance. Note that returned node may still be the same as this node
+  /// instance. Returned node may still be the same as this node
   /// if provided [index] is `0`.
   LeafNode isolate(int index, int length) {
     assert(
@@ -88,13 +97,13 @@ abstract class LeafNode extends Node
         'Actual node length: ${this.length}.');
     // Since `index < this.length` (guarded by assert) below line
     // always returns a new node.
-    final target = splitAt(index);
+    final target = splitAt(index)!;
     target.splitAt(length);
     return target;
   }
 
   /// Formats this node and optimizes it with adjacent leaf nodes if needed.
-  void formatAndOptimize(NotusStyle style) {
+  void formatAndOptimize(NotusStyle? style) {
     if (style != null && style.isNotEmpty) {
       applyStyle(style);
     }
@@ -103,23 +112,8 @@ abstract class LeafNode extends Node
 
   @override
   void applyStyle(NotusStyle value) {
-    assert(value != null && (value.isInline || value.isEmpty),
+    assert(value.isInline || value.isEmpty,
         'Style cannot be applied to this leaf node: $value');
-    assert(() {
-      if (value.contains(NotusAttribute.embed)) {
-        if (value.get(NotusAttribute.embed) == NotusAttribute.embed.unset) {
-          throw 'Unsetting embed attribute is not allowed. '
-              'This operation means that the embed itself must be deleted from the document. '
-              'Make sure there is FormatEmbedsRule in your heuristics registry, '
-              'which is responsible for handling this scenario.';
-        }
-        if (this is! EmbedNode) {
-          throw 'Embed style can only be applied to an EmbedNode.';
-        }
-      }
-      return true;
-    }());
-
     super.applyStyle(value);
   }
 
@@ -127,31 +121,36 @@ abstract class LeafNode extends Node
   LineNode get parent => super.parent as LineNode;
 
   @override
-  int get length => _value.length;
-
-  @override
-  Delta toDelta() {
-    return Delta()..insert(_value, style.toJson());
+  int get length {
+    if (_value is String) {
+      return (_value as String).length;
+    }
+    return 1; // embed objects have length of 1.
   }
 
   @override
-  String toPlainText() => _value;
+  Delta toDelta() {
+    final data = _value is EmbeddableObject
+        ? (_value as EmbeddableObject).toJson()
+        : _value;
+    return Delta()..insert(data, style.toJson());
+  }
 
   @override
-  void insert(int index, String value, NotusStyle style) {
-    assert(index >= 0 && (index <= length), 'Index: $index, Length: $length.');
-    assert(value.isNotEmpty);
-    final node = LeafNode(value);
+  void insert(int index, Object data, NotusStyle? style) {
+    assert(index >= 0 && (index <= length),
+        'Index out of bounds. Must be between 0 and $length, but got $index.');
+    final node = LeafNode(data);
     if (index == length) {
       insertAfter(node);
     } else {
-      splitAt(index).insertBefore(node);
+      splitAt(index)!.insertBefore(node);
     }
     node.formatAndOptimize(style);
   }
 
   @override
-  void retain(int index, int length, NotusStyle style) {
+  void retain(int index, int length, NotusStyle? style) {
     if (style == null) return;
 
     final local = math.min(this.length - index, length);
@@ -160,7 +159,7 @@ abstract class LeafNode extends Node
     final remaining = length - local;
     if (remaining > 0) {
       assert(node.next != null);
-      node.next.retain(0, remaining, style);
+      node.next!.retain(0, remaining, style);
     }
     // Optimize at the very end
     node.formatAndOptimize(style);
@@ -180,7 +179,7 @@ abstract class LeafNode extends Node
     final remaining = length - local;
     if (remaining > 0) {
       assert(actualNext != null);
-      actualNext.delete(0, remaining);
+      actualNext!.delete(0, remaining);
     }
 
     if (needsOptimize != null) needsOptimize.optimize();
@@ -197,19 +196,30 @@ abstract class LeafNode extends Node
   /// the same style.
   @override
   void optimize() {
-    var node = this;
-    if (!node.isFirst) {
-      LeafNode mergeWith = node.previous;
+    if (this is EmbedNode) {
+      // Embed nodes cannot be merged with text nor other embeds (in fact,
+      // there could be no two adjacent embeds on the same line since an
+      // embed occupies an entire line).
+      return;
+    }
+
+    // This is a text node and it can only be merged with other text nodes.
+
+    var node = this as TextNode;
+    if (!node.isFirst && node.previous is TextNode) {
+      var mergeWith = node.previous as TextNode;
       if (mergeWith.style == node.style) {
-        mergeWith._value += node.value;
+        final combinedValue = mergeWith.value + node.value;
+        mergeWith._value = combinedValue;
         node.unlink();
         node = mergeWith;
       }
     }
-    if (!node.isLast) {
-      LeafNode mergeWith = node.next;
+    if (!node.isLast && node.next is TextNode) {
+      var mergeWith = node.next as TextNode;
       if (mergeWith.style == node.style) {
-        node._value += mergeWith._value;
+        final combinedValue = node.value + mergeWith.value;
+        node._value = combinedValue;
         mergeWith.unlink();
       }
     }
@@ -225,30 +235,45 @@ abstract class LeafNode extends Node
 ///
 /// See also:
 ///
+///   * [EmbedNode], a leaf node representing an embeddable object.
 ///   * [LineNode], a node representing a line of text.
 ///   * [BlockNode], a node representing a group of lines.
 class TextNode extends LeafNode {
-  TextNode([String content = '']) : super._(content);
-}
+  TextNode([String text = ''])
+      : assert(!text.contains('\n')),
+        super._(text);
 
-final kZeroWidthSpace = String.fromCharCode(0x200b);
+  @override
+  String get value => _value as String;
+
+  @override
+  String toPlainText() => value;
+}
 
 /// An embed node inside of a line in a Notus document.
 ///
 /// Embed node is a leaf node similar to [TextNode]. It represents an
-/// arbitrary piece of non-text content embedded into a document, such as,
+/// arbitrary piece of non-textual content embedded into a document, such as,
 /// image, horizontal rule, video, or any other object with defined structure,
-/// like tweet, for instance.
+/// like a tweet, for instance.
 ///
 /// Embed node's length is always `1` character and it is represented with
-/// zero-width space in the document text.
+/// unicode object replacement character in the document text.
 ///
 /// Any inline style can be applied to an embed, however this does not
 /// necessarily mean the embed will look according to that style. For instance,
 /// applying "bold" style to an image gives no effect, while adding a "link" to
 /// an image actually makes the image react to user's action.
 class EmbedNode extends LeafNode {
-  static final kPlainTextPlaceholder = String.fromCharCode(0x200b);
+  static final kObjectReplacementCharacter = '\uFFFC';
 
-  EmbedNode() : super._(kPlainTextPlaceholder);
+  EmbedNode(EmbeddableObject object) : super._(object);
+
+  @override
+  EmbeddableObject get value => super.value as EmbeddableObject;
+
+  // Embed nodes are represented as unicode object replacement character in
+  // plain text.
+  @override
+  String toPlainText() => kObjectReplacementCharacter;
 }
