@@ -12,44 +12,49 @@ abstract class FormatRule {
 
   /// Applies heuristic rule to a retain (format) operation on a [document] and
   /// returns resulting [Delta].
-  Delta apply(Delta document, int index, int length, NotusAttribute attribute);
+  Delta? apply(Delta document, int index, int length, NotusAttribute attribute);
 }
 
 /// Produces Delta with line-level attributes applied strictly to
-/// line-break characters.
+/// newline characters.
 class ResolveLineFormatRule extends FormatRule {
   const ResolveLineFormatRule() : super();
 
   @override
-  Delta apply(Delta document, int index, int length, NotusAttribute attribute) {
+  Delta? apply(
+      Delta document, int index, int length, NotusAttribute attribute) {
     if (attribute.scope != NotusAttributeScope.line) return null;
 
     var result = Delta()..retain(index);
     final iter = DeltaIterator(document);
     iter.skip(index);
 
-    // Apply line styles to all line-break characters within range of this
+    // Apply line styles to all newline characters within range of this
     // retain operation.
     var current = 0;
     while (current < length && iter.hasNext) {
       final op = iter.next(length - current);
-      if (op.data.contains('\n')) {
-        final delta = _applyAttribute(op.data, attribute);
+      final opText = op.data is String ? op.data as String : '';
+      if (opText.contains('\n')) {
+        final delta = _applyAttribute(op.data as String, attribute);
         result = result.concat(delta);
       } else {
         result.retain(op.length);
       }
       current += op.length;
     }
-    // And include extra line-break after retain
+    // And include extra newline after retain
     while (iter.hasNext) {
       final op = iter.next();
-      final lf = op.data.indexOf('\n');
+      final opText = op.data is String ? op.data as String : '';
+      final lf = opText.indexOf('\n');
       if (lf == -1) {
-        result..retain(op.length);
+        result.retain(op.length);
         continue;
       }
-      result..retain(lf)..retain(1, attribute.toJson());
+      result
+        ..retain(lf)
+        ..retain(1, attribute.toJson());
       break;
     }
     return result;
@@ -60,7 +65,9 @@ class ResolveLineFormatRule extends FormatRule {
     var offset = 0;
     var lf = text.indexOf('\n');
     while (lf >= 0) {
-      result..retain(lf - offset)..retain(1, attribute.toJson());
+      result
+        ..retain(lf - offset)
+        ..retain(1, attribute.toJson());
       offset = lf + 1;
       lf = text.indexOf('\n', offset);
     }
@@ -71,30 +78,34 @@ class ResolveLineFormatRule extends FormatRule {
 }
 
 /// Produces Delta with inline-level attributes applied too all characters
-/// except line-breaks.
+/// except newlines.
 class ResolveInlineFormatRule extends FormatRule {
   const ResolveInlineFormatRule();
 
   @override
-  Delta apply(Delta document, int index, int length, NotusAttribute attribute) {
+  Delta? apply(
+      Delta document, int index, int length, NotusAttribute attribute) {
     if (attribute.scope != NotusAttributeScope.inline) return null;
 
     final result = Delta()..retain(index);
     final iter = DeltaIterator(document);
     iter.skip(index);
 
-    // Apply inline styles to all non-line-break characters within range of this
+    // Apply inline styles to all non-newline characters within range of this
     // retain operation.
     var current = 0;
     while (current < length && iter.hasNext) {
       final op = iter.next(length - current);
-      var lf = op.data.indexOf('\n');
+      final opText = op.data is String ? op.data as String : '';
+      var lf = opText.indexOf('\n');
       if (lf != -1) {
         var pos = 0;
         while (lf != -1) {
-          result..retain(lf - pos, attribute.toJson())..retain(1);
+          result
+            ..retain(lf - pos, attribute.toJson())
+            ..retain(1);
           pos = lf + 1;
-          lf = op.data.indexOf('\n', pos);
+          lf = opText.indexOf('\n', pos);
         }
         if (pos < op.length) result.retain(op.length - pos, attribute.toJson());
       } else {
@@ -112,7 +123,8 @@ class FormatLinkAtCaretPositionRule extends FormatRule {
   const FormatLinkAtCaretPositionRule();
 
   @override
-  Delta apply(Delta document, int index, int length, NotusAttribute attribute) {
+  Delta? apply(
+      Delta document, int index, int length, NotusAttribute attribute) {
     if (attribute.key != NotusAttribute.link.key) return null;
     // If user selection is not collapsed we let it fallback to default rule
     // which simply applies the attribute to selected range.
@@ -132,87 +144,17 @@ class FormatLinkAtCaretPositionRule extends FormatRule {
       startIndex -= before.length;
       retain = before.length;
     }
-    if (after != null && after.hasAttribute(attribute.key)) {
+    if (after.hasAttribute(attribute.key)) {
       retain += after.length;
     }
     // There is no link-styled text around `index` position so it becomes a
     // no-op action.
     if (retain == 0) return null;
 
-    result..retain(startIndex)..retain(retain, attribute.toJson());
+    result
+      ..retain(startIndex)
+      ..retain(retain, attribute.toJson());
 
     return result;
-  }
-}
-
-/// Handles all format operations which manipulate embeds.
-class FormatEmbedsRule extends FormatRule {
-  const FormatEmbedsRule();
-
-  @override
-  Delta apply(Delta document, int index, int length, NotusAttribute attribute) {
-    // We are only interested in embed attributes
-    if (attribute is! EmbedAttribute) return null;
-    EmbedAttribute embed = attribute;
-
-    if (length == 1 && embed.isUnset) {
-      // Remove the embed.
-      return Delta()
-        ..retain(index)
-        ..delete(length);
-    } else {
-      // If length is 0 we treat it as an insert at specified [index].
-      // If length is non-zero we treat it as a replace of selected range
-      // with the embed.
-      assert(!embed.isUnset);
-      return _insertEmbed(document, index, length, embed);
-    }
-  }
-
-  Delta _insertEmbed(
-      Delta document, int index, int length, EmbedAttribute embed) {
-    final result = Delta()..retain(index);
-    final iter = DeltaIterator(document);
-    final previous = iter.skip(index);
-    iter.skip(length); // ignore deleted part.
-    final target = iter.next();
-
-    // Check if [index] is on an empty line already.
-    final isNewlineBefore = previous == null || previous.data.endsWith('\n');
-    final isNewlineAfter = target.data.startsWith('\n');
-    final isOnEmptyLine = isNewlineBefore && isNewlineAfter;
-    if (isOnEmptyLine) {
-      return result..insert(EmbedNode.kPlainTextPlaceholder, embed.toJson());
-    }
-    // We are on a non-empty line, split it (preserving style if needed)
-    // and insert our embed.
-    final lineStyle = _getLineStyle(iter, target);
-    if (!isNewlineBefore) {
-      result..insert('\n', lineStyle);
-    }
-    result..insert(EmbedNode.kPlainTextPlaceholder, embed.toJson());
-    if (!isNewlineAfter) {
-      result..insert('\n');
-    }
-    result.delete(length);
-    return result;
-  }
-
-  Map<String, dynamic> _getLineStyle(
-      DeltaIterator iterator, Operation current) {
-    if (current.data.contains('\n')) {
-      return current.attributes;
-    }
-    // Continue looking for line-break.
-    Map<String, dynamic> attributes;
-    while (iterator.hasNext) {
-      final op = iterator.next();
-      final lf = op.data.indexOf('\n');
-      if (lf >= 0) {
-        attributes = op.attributes;
-        break;
-      }
-    }
-    return attributes;
   }
 }
