@@ -3,12 +3,10 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:notus/notus.dart';
-import 'package:quill_delta/quill_delta.dart';
 import 'package:test/test.dart';
 
 NotusDocument dartconfDoc() {
-  final delta = Delta()..insert('DartConf\nLos Angeles\n');
-  return NotusDocument.fromDelta(delta);
+  return NotusDocument()..insert(0, 'DartConf\nLos Angeles');
 }
 
 final ul = NotusAttribute.ul.toJson();
