@@ -1,53 +1,144 @@
+<<<<<<< HEAD
 // Copyright (c) 2018, the Zefyr project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 import 'dart:math' as math;
 
 import 'package:flutter/widgets.dart';
+import 'dart:math' as math;
+import 'package:flutter/cupertino.dart';
 import 'package:notus/notus.dart';
 import 'package:quill_delta/quill_delta.dart';
 import 'package:zefyr/util.dart';
 
-const TextSelection _kZeroSelection = TextSelection.collapsed(
-  offset: 0,
-  affinity: TextAffinity.upstream,
-);
+/// List of style keys which can be toggled for insertion
+List<String> _insertionToggleableStyleKeys = [
+  NotusAttribute.bold.key,
+  NotusAttribute.italic.key,
+  NotusAttribute.underline.key,
+  NotusAttribute.strikethrough.key,
+];
 
-/// Owner of focus.
-enum FocusOwner {
-  /// Current owner is the editor.
-  editor,
-
-  /// Current owner is the toolbar.
-  toolbar,
-
-  /// No focus owner.
-  none,
-}
-
-/// Controls instance of [ZefyrEditor].
 class ZefyrController extends ChangeNotifier {
-  ZefyrController(NotusDocument document)
-      : assert(document != null),
-        document = document;
+  ZefyrController([NotusDocument? document])
+      : document = document ?? NotusDocument(),
+        _selection = TextSelection.collapsed(offset: 0);
 
-  /// Zefyr document managed by this controller.
+  /// Document managed by this controller.
   final NotusDocument document;
 
   /// Currently selected text within the [document].
   TextSelection get selection => _selection;
-  TextSelection _selection = _kZeroSelection;
-
-  ChangeSource _lastChangeSource;
-
-  /// Source of the last text or selection change.
-  ChangeSource get lastChangeSource => _lastChangeSource;
+  TextSelection _selection;
 
   /// Store any styles attribute that got toggled by the tap of a button
   /// and that has not been applied yet.
-  /// It gets reseted after each format action within the [document].
+  /// It gets reset after each format action within the [document].
   NotusStyle get toggledStyles => _toggledStyles;
   NotusStyle _toggledStyles = NotusStyle();
+
+  /// Returns style of specified text range.
+  ///
+  /// If nothing is selected but we've toggled an attribute,
+  /// we also merge those in our style before returning.
+  NotusStyle getSelectionStyle() {
+    final start = _selection.start;
+    final length = _selection.end - start;
+    var lineStyle = document.collectStyle(start, length);
+
+    lineStyle = lineStyle.mergeAll(toggledStyles);
+
+    return lineStyle;
+  }
+
+  /// Replaces [length] characters in the document starting at [index] with
+  /// provided [text].
+  ///
+  /// Resulting change is registered as produced by user action, e.g.
+  /// using [ChangeSource.local].
+  ///
+  /// It also applies the toggledStyle if needed. And then it resets it
+  /// in any cases as we don't want to keep it except on inserts.
+  ///
+  /// Optionally updates selection if provided.
+  void replaceText(int index, int length, Object data,
+      {TextSelection? selection}) {
+    assert(data is String || data is EmbeddableObject);
+    Delta? delta;
+
+    final isDataNotEmpty = data is String ? data.isNotEmpty : true;
+    if (length > 0 || isDataNotEmpty) {
+      delta = document.replace(index, length, data);
+      // If the delta is an insert operation and we have toggled
+      // some styles, then apply those styles to the inserted text.
+      if (toggledStyles.isNotEmpty &&
+          delta.isNotEmpty &&
+          delta.length <= 2 && // covers single insert and a retain+insert
+          delta.last.isInsert) {
+        final dataLength = data is String ? data.length : 1;
+        final retainDelta = Delta()
+          ..retain(index)
+          ..retain(dataLength, toggledStyles.toJson());
+        document.compose(retainDelta, ChangeSource.local);
+      }
+    }
+
+    // Always reset it after any user action, even if it has not been applied.
+    _toggledStyles = NotusStyle();
+
+    if (selection != null) {
+      if (delta == null) {
+        _updateSelectionSilent(selection, source: ChangeSource.local);
+      } else {
+        // need to transform selection position in case actual delta
+        // is different from user's version (in deletes and inserts).
+        final user = Delta()
+          ..retain(index)
+          ..insert(data)
+          ..delete(length);
+        var positionDelta = getPositionDelta(user, delta);
+        _updateSelectionSilent(
+          selection.copyWith(
+            baseOffset: selection.baseOffset + positionDelta,
+            extentOffset: selection.extentOffset + positionDelta,
+          ),
+          source: ChangeSource.local,
+        );
+      }
+    }
+//    _lastChangeSource = ChangeSource.local;
+    notifyListeners();
+  }
+
+  void formatText(int index, int length, NotusAttribute attribute) {
+    final change = document.format(index, length, attribute);
+    // _lastChangeSource = ChangeSource.local;
+    final source = ChangeSource.local;
+
+    if (length == 0 && _insertionToggleableStyleKeys.contains(attribute.key)) {
+      // Add the attribute to our toggledStyle. It will be used later upon insertion.
+      _toggledStyles = toggledStyles.put(attribute);
+    }
+
+    // Transform selection against the composed change and give priority to
+    // the change. This is needed in cases when format operation actually
+    // inserts data into the document (e.g. embeds).
+    final base = change.transformPosition(_selection.baseOffset);
+    final extent = change.transformPosition(_selection.extentOffset);
+    final adjustedSelection =
+        _selection.copyWith(baseOffset: base, extentOffset: extent);
+    if (_selection != adjustedSelection) {
+      _updateSelectionSilent(adjustedSelection, source: source);
+    }
+    notifyListeners();
+  }
+
+  /// Formats current selection with [attribute].
+  void formatSelection(NotusAttribute attribute) {
+    final index = _selection.start;
+    final length = _selection.end - index;
+    formatText(index, length, attribute);
+  }
 
   /// Updates selection with specified [value].
   ///
@@ -58,21 +149,6 @@ class ZefyrController extends ChangeNotifier {
     notifyListeners();
   }
 
-  // Updates selection without triggering notifications to listeners.
-  void _updateSelectionSilent(TextSelection value,
-      {ChangeSource source = ChangeSource.remote}) {
-    assert(value != null && source != null);
-    _selection = value;
-    _lastChangeSource = source;
-    _ensureSelectionBeforeLastBreak();
-  }
-
-  @override
-  void dispose() {
-    document.close();
-    super.dispose();
-  }
-
   /// Composes [change] into document managed by this controller.
   ///
   /// This method does not apply any adjustments or heuristic rules to
@@ -81,7 +157,7 @@ class ZefyrController extends ChangeNotifier {
   ///
   /// If composing this change fails then this method throws [ComposeError].
   void compose(Delta change,
-      {TextSelection selection, ChangeSource source = ChangeSource.remote}) {
+      {TextSelection? selection, ChangeSource source = ChangeSource.remote}) {
     if (change.isNotEmpty) {
       document.compose(change, source);
     }
@@ -99,126 +175,41 @@ class ZefyrController extends ChangeNotifier {
         _updateSelectionSilent(selection, source: source);
       }
     }
-    _lastChangeSource = source;
     notifyListeners();
   }
 
-  /// Replaces [length] characters in the document starting at [index] with
-  /// provided [text].
-  ///
-  /// Resulting change is registered as produced by user action, e.g.
-  /// using [ChangeSource.local].
-  ///
-  /// It also applies the toggledStyle if needed. And then it resets it
-  /// in any cases as we don't want to keep it except on inserts.
-  ///
-  /// Optionally updates selection if provided.
-  void replaceText(int index, int length, String text,
-      {TextSelection selection}) {
-    Delta delta;
-
-    if (length > 0 || text.isNotEmpty) {
-      delta = document.replace(index, length, text);
-      // If the delta is a classical insert operation and we have toggled
-      // some style, then we apply it to our document.
-      if (delta != null &&
-          toggledStyles.isNotEmpty &&
-          delta.isNotEmpty &&
-          delta.length <= 2 &&
-          delta.last.isInsert) {
-        // Apply it.
-        final retainDelta = Delta()
-          ..retain(index)
-          ..retain(text.length, toggledStyles.toJson());
-        document.compose(retainDelta, ChangeSource.local);
-      }
-    }
-
-    // Always reset it after any user action, even if it has not been applied.
-    _toggledStyles = NotusStyle();
-
-    if (selection != null) {
-      if (delta == null) {
-        _updateSelectionSilent(selection, source: ChangeSource.local);
-      } else {
-        // need to transform selection position in case actual delta
-        // is different from user's version (in deletes and inserts).
-        final user = Delta()
-          ..retain(index)
-          ..insert(text)
-          ..delete(length);
-        var positionDelta = getPositionDelta(user, delta);
-        _updateSelectionSilent(
-          selection.copyWith(
-            baseOffset: selection.baseOffset + positionDelta,
-            extentOffset: selection.extentOffset + positionDelta,
-          ),
-          source: ChangeSource.local,
-        );
-      }
-    }
-    _lastChangeSource = ChangeSource.local;
-    notifyListeners();
+  @override
+  void dispose() {
+    document.close();
+    super.dispose();
   }
 
-  void formatText(int index, int length, NotusAttribute attribute) {
-    final change = document.format(index, length, attribute);
-    _lastChangeSource = ChangeSource.local;
-
-    if (length == 0 &&
-        (attribute.key == NotusAttribute.bold.key ||
-            attribute.key == NotusAttribute.italic.key ||
-            attribute.key == NotusAttribute.underline.key)) {
-      // Add the attribute to our toggledStyle. It will be used later upon insertion.
-      _toggledStyles = toggledStyles.put(attribute);
-    }
-
-    // Transform selection against the composed change and give priority to
-    // the change. This is needed in cases when format operation actually
-    // inserts data into the document (e.g. embeds).
-    final base = change.transformPosition(_selection.baseOffset);
-    final extent = change.transformPosition(_selection.extentOffset);
-    final adjustedSelection =
-        _selection.copyWith(baseOffset: base, extentOffset: extent);
-    if (_selection != adjustedSelection) {
-      _updateSelectionSilent(adjustedSelection, source: _lastChangeSource);
-    }
-    notifyListeners();
+  /// Updates selection without triggering notifications to listeners.
+  void _updateSelectionSilent(TextSelection value,
+      {ChangeSource source = ChangeSource.remote}) {
+    _selection = value;
+//    _lastChangeSource = source;
+    _ensureSelectionBeforeLastBreak();
   }
 
-  /// Formats current selection with [attribute].
-  void formatSelection(NotusAttribute attribute) {
-    final index = _selection.start;
-    final length = _selection.end - index;
-    formatText(index, length, attribute);
-  }
-
-  /// Returns style of specified text range.
-  ///
-  /// If nothing is selected but we've toggled an attribute,
-  ///  we also merge those in our style before returning.
-  NotusStyle getSelectionStyle() {
-    final start = _selection.start;
-    final length = _selection.end - start;
-    var lineStyle = document.collectStyle(start, length);
-
-    lineStyle = lineStyle.mergeAll(toggledStyles);
-
-    return lineStyle;
-  }
-
-  TextEditingValue get plainTextEditingValue {
-    return TextEditingValue(
-      text: document.toPlainText(),
-      selection: selection,
-      composing: TextRange.collapsed(-1),
-    );
-  }
-
+  // Ensures that selection does not include last line break which
+  // prevents deletion of the last line in the document.
+  // This is required by Notus document model.
   void _ensureSelectionBeforeLastBreak() {
     final end = document.length - 1;
     final base = math.min(_selection.baseOffset, end);
     final extent = math.min(_selection.extentOffset, end);
     _selection = _selection.copyWith(baseOffset: base, extentOffset: extent);
   }
+<<<<<<< HEAD
+=======
+
+  TextEditingValue get plainTextEditingValue {
+    return TextEditingValue(
+      text: document.toPlainText(),
+      selection: selection,
+      composing: TextRange.empty,
+    );
+  }
+>>>>>>> upstream/1.0-dev
 }
