@@ -26,7 +26,8 @@ int getPositionDelta(Delta user, Delta actual) {
     } else if (userOp.isDelete && actualOp.isRetain) {
       diff += userOp.length;
     } else if (userOp.isRetain && actualOp.isInsert) {
-      if (actualOp.data.startsWith('\n')) {
+      final opText = actualOp.data is String ? actualOp.data as String : '';
+      if (opText.startsWith('\n')) {
         // At this point user input reached its end (retain). If a heuristic
         // rule inserts a new line we should keep cursor on it's original position.
         continue;
