@@ -9,22 +9,11 @@ library zefyr;
 
 export 'package:notus/notus.dart';
 
-export 'src/widgets/buttons.dart' hide HeadingButton, LinkButton;
-export 'src/widgets/code.dart';
-export 'src/widgets/common.dart';
+export 'src/rendering/editor.dart';
 export 'src/widgets/controller.dart';
-export 'src/widgets/editable_text.dart';
+export 'src/widgets/cursor.dart';
 export 'src/widgets/editor.dart';
+export 'src/widgets/editor_toolbar.dart';
 export 'src/widgets/field.dart';
-export 'src/widgets/horizontal_rule.dart';
-export 'src/widgets/image.dart';
-export 'src/widgets/list.dart';
-export 'src/widgets/mode.dart';
-export 'src/widgets/paragraph.dart';
-export 'src/widgets/quote.dart';
-export 'src/widgets/scaffold.dart';
-export 'src/widgets/scope.dart' hide ZefyrScopeAccess;
-export 'src/widgets/selection.dart' hide SelectionHandleDriver;
+export 'src/widgets/text_line.dart';
 export 'src/widgets/theme.dart';
-export 'src/widgets/toolbar.dart';
-export 'src/widgets/view.dart';
