@@ -3,12 +3,11 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:zefyr/zefyr.dart';
 
 import '../testing.dart';
 
 void main() {
-  group('$ZefyrEditableText', () {
+  group('ZefyrEditableText', () {
     testWidgets('user input', (tester) async {
       final editor = EditorSandBox(tester: tester);
       await editor.pumpAndTap();
